Update README.md
"I have successfully developed an online quiz application as part of a task for Digital Bhem company. This application allows users to take quizzes on various topics and receive immediate feedback on their performance. It is built using C++ and features a user-friendly interface. The project demonstrates my proficiency in C++ programming and software development skills."